+++ src/main/java/de/retest/recheck/configuration/ProjectConfigurationUtil.java
@@ -12,8 +12,4 @@ public class ProjectConfigurationUtil {
 	public static Optional<Path> findProjectConfigurationFolder() {
 		return ProjectRootFinderUtil.getProjectRoot().map( path -> path.resolve( RETEST_FOLDER_NAME ) );
 	}
-
-	public static Optional<Path> findProjectConfigurationFolder( final Path basePath ) {
-		return ProjectRootFinderUtil.getProjectRoot( basePath ).map( path -> path.resolve( RETEST_FOLDER_NAME ) );
-	}
 }