+++ src/main/java/de/retest/recheck/ignore/RecheckIgnoreUtil.java
@@ -1,15 +1,17 @@
 package de.retest.recheck.ignore;
 
+import static de.retest.recheck.RecheckProperties.RETEST_FOLDER_NAME;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import de.retest.recheck.configuration.ProjectConfiguration;
 import de.retest.recheck.configuration.ProjectConfigurationUtil;
 import de.retest.recheck.review.GlobalIgnoreApplier;
 import de.retest.recheck.review.counter.NopCounter;
@@ -21,34 +23,30 @@ public class RecheckIgnoreUtil {
 
 	private RecheckIgnoreUtil() {}
 
-	public static Optional<Path> getIgnoreFile() {
-		return getRetestFile( ProjectConfiguration.RECHECK_IGNORE );
-	}
-
-	public static Optional<Path> getIgnoreRuleFile() {
-		return getRetestFile( ProjectConfiguration.RECHECK_IGNORE_JSRULES );
-	}
-
-	private static Optional<Path> getRetestFile( final String filename ) {
+	public static Optional<Path> getProjectIgnoreFile( final String filename ) {
 		final Optional<Path> projectConfigurationFolder = ProjectConfigurationUtil.findProjectConfigurationFolder();
 		return projectConfigurationFolder.map( p -> p.resolve( filename ) );
 	}
 
-	public static GlobalIgnoreApplier loadRecheckIgnore() {
-		return loadRecheckSuiteIgnore( new LoadFilterWorker( NopCounter.getInstance() ) );
+	public static Path getUserIgnoreFile( final String filename ) {
+		return Paths.get( System.getProperty( "user.home" ) ).resolve( RETEST_FOLDER_NAME ).resolve( filename );
 	}
 
-	public static GlobalIgnoreApplier loadRecheckSuiteIgnore( final File ignoreFilesBasePath ) {
-		return loadRecheckSuiteIgnore( new LoadFilterWorker( NopCounter.getInstance(), ignoreFilesBasePath.toPath() ) );
+	public static Path getSuiteIgnoreFile( final String filename, final Path basePath ) {
+		return basePath.resolve( filename );
+	}
+
+	public static GlobalIgnoreApplier loadRecheckIgnore( final File suiteIgnorePath ) {
+		return loadRecheckSuiteIgnore( new LoadFilterWorker( NopCounter.getInstance(), suiteIgnorePath.toPath() ) );
 	}
 
 	private static GlobalIgnoreApplier loadRecheckSuiteIgnore( final LoadFilterWorker loadFilterWorker ) {
 		try {
 			return loadFilterWorker.load();
 		} catch ( final NoSuchFileException | FileNotFoundException e ) {
-			logger.debug( "Ignoring missing suite ignore file." );
+			logger.debug( "Ignoring missing suite or user ignore file." );
 		} catch ( final Exception e ) {
-			logger.error( "Exception loading suite ignore file.", e );
+			logger.error( "Exception loading suite or user ignore file.", e );
 		}
 		return GlobalIgnoreApplier.create( NopCounter.getInstance() );
 	}